設定系画面修正 リファクタ

+++ app/src/main/java/com/example/musicdictionaryandroid/model/usecase/ArtistUseCaseImp.kt
@@ -1,7 +1,8 @@
 package com.example.musicdictionaryandroid.model.usecase
 
+import androidx.lifecycle.LiveData
+import com.example.musicdictionaryandroid.model.entity.Artist
 import com.example.musicdictionaryandroid.model.entity.ArtistsForm
-import com.example.musicdictionaryandroid.model.entity.CallBackData
 import com.example.musicdictionaryandroid.model.repository.*
 import com.example.musicdictionaryandroid.model.util.Result
 import java.util.ArrayList
@@ -57,10 +58,10 @@ class ArtistUseCaseImp(
           return withContext(Dispatchers.IO) {
               try {
                   val result = apiRepository.getArtistsByEmail(email)
-                  result.body()?.let{
+                  result.body()?.let {
                       PreferenceRepositoryImp.setFavorite(it.size)
                       dataBaseRepository.updateAll(it)
-                  }?: run { PreferenceRepositoryImp.setFavorite(0) }
+                  } ?: run { PreferenceRepositoryImp.setFavorite(0) }
                   return@withContext Result.Success(result.body())
               } catch (e: Exception) {
                   return@withContext Result.Success(dataBaseRepository.getArtistAll())
@@ -69,43 +70,77 @@ class ArtistUseCaseImp(
     }
 
     // アーティスト登録
-    override suspend fun addArtist(artist: ArtistsForm, email: String): Result<ArtistsForm?> {
+    override suspend fun addArtist(artist: Artist, email: String): Result<ArtistsForm?> {
         return withContext(Dispatchers.IO) {
             try {
-                val result = apiRepository.addArtist(artist, email)
-                dataBaseRepository.addArtist(result.body()!!)
-                return@withContext Result.Success(result.body())
+                val artistsFrom = ArtistsForm(
+                    artist.name!!,
+                    artist.gender!!,
+                    artist.voice!!,
+                    artist.length!!,
+                    artist.lyrics!!,
+                    artist.genre1!!,
+                    artist.genre2!!
+                )
+                val result = apiRepository.addArtist(artistsFrom, email)
+                if (result.body() != null) {
+                    dataBaseRepository.addArtist(result.body()!!)
+                    return@withContext Result.Success(result.body())
+                } else {
+                    @Suppress("UNREACHABLE_CODE")
+                    return@withContext Result.Error(throw Exception("サーバーエラー: 更新できませんでした"))
+                }
             } catch (e: Exception) {
                 return@withContext Result.Error(e)
             }
         }
     }
     // アーティスト更新
-    override suspend fun updateArtist(artist: ArtistsForm, email: String): Result<ArtistsForm?> {
+    override suspend fun updateArtist(artist: Artist, email: String): Result<ArtistsForm?> {
         return withContext(Dispatchers.IO) {
             try {
-                val result = apiRepository.updateArtist(artist, email)
-                dataBaseRepository.deleteAll()
-                dataBaseRepository.updateArtist(artist)
-                return@withContext Result.Success(result.body())
+                val artistsFrom = ArtistsForm(
+                    artist.name!!,
+                    artist.gender!!,
+                    artist.voice!!,
+                    artist.length!!,
+                    artist.lyrics!!,
+                    artist.genre1!!,
+                    artist.genre2!!
+                )
+                val result = apiRepository.updateArtist(artistsFrom, email)
+                if (result.body() != null) {
+                    dataBaseRepository.deleteAll()
+                    dataBaseRepository.updateArtist(result.body()!!)
+                    return@withContext Result.Success(result.body())
+                } else {
+                    @Suppress("UNREACHABLE_CODE")
+                    return@withContext Result.Error(throw Exception("サーバーエラー: 更新できませんでした"))
+                }
             } catch (e: Exception) {
                 return@withContext Result.Error(e)
             }
         }
     }
     // アーティスト削除
-    override suspend fun deleteArtist(name: String, email: String): Result<CallBackData?> {
+    override suspend fun deleteArtist(name: String, email: String): Result<List<ArtistsForm>> {
         return withContext(Dispatchers.IO) {
             try {
-                val result = apiRepository.deleteArtist(name, email)
+                apiRepository.deleteArtist(name, email)
+                dataBaseRepository.deleteArtist(name)
                 val size = PreferenceRepositoryImp.getFavorite()
                 PreferenceRepositoryImp.setFavorite(size - 1)
-                dataBaseRepository.deleteArtist(name)
-                return@withContext Result.Success(result.body())
+                val artist = dataBaseRepository.getArtistAll()
+                return@withContext Result.Success(artist)
             } catch (e: Exception) {
                 return@withContext Result.Error(e)
             }
         }
+    }
+
+    // アーティストリスト取得
+    override fun getArtistList(): LiveData<List<Artist>> {
+        return dataBaseRepository.getArtistList()
     }
 
     //endregion
